Fix S1118, S1905, S3903, S2190: utility constructors, redundant casts, namespaces, infinite loop
- S1118 (19): add private constructors to static utility classes (incl. Program entry-point classes).

- S1905 (16): remove redundant casts (object?/float/ulong), NOSONAR where casts are required for type inference.

- S3903 (16): move explicit-class files into file-scoped namespaces (6 files, incl. StartupObject update); NOSONAR on trailing types after top-level statements (EF demos).

- S2190 (1): document intentional continuous-profiling loop.

Verified: SharpCoreDB.CI.slnf + affected projects build with 0 errors; all 1397 SharpCoreDB.Tests pass.

+++ src/SharpCoreDB/Storage/SingleFileStorageProvider.cs
@@ -1463,7 +1463,7 @@ public sealed class SingleFileStorageProvider : IStorageProvider
         // Without this, FreeSpaceManager starts with _totalPages=0 and AllocatePages
         // returns offset 0 (the SCDB header page!). Data block writes then overwrite the
         // file header with table data (e.g. "SFT1" magic), corrupting the file.
-        var reservedPages = (ulong)(totalMetadataSize / (ulong)options.PageSize);
+        var reservedPages = totalMetadataSize / (ulong)options.PageSize;
         header.AllocatedPages = reservedPages;
 
         var fsmHeader = new FreeSpaceMapHeader
